+++ src/controllers/plugins/LuaAPI.hpp
@@ -88,6 +88,7 @@ sol::table toTable(lua_State *L, const CompletionEvent &ev);
  * @includefile controllers/plugins/api/HTTPRequest.hpp
  * @includefile controllers/plugins/api/Message.hpp
  * @includefile controllers/plugins/api/WebSocket.hpp
+ * @includefile controllers/plugins/api/WindowManager.hpp
  * @includefile common/network/NetworkCommon.hpp
  */