+++ src/widgets/splits/SplitContainer.hpp
@@ -82,6 +82,9 @@ public:
         Node();
         Node(Split *_split, Node *_parent);
 
+        /**
+         * @exposeenum c2.SplitContainerNodeType
+         */
         enum class Type {
             EmptyRoot,
             Split,